requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.4.0
requests>=2.31
gunicorn
dj-database-url
dnspython

google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0